requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
packaging==23.2
pydantic==2.6.2
pydantic_core==2.16.3
python-dotenv==1.0.1
replicate==0.24.0
sniffio==1.3.0
typing_extensions==4.9.0
